requirements.txt
chromadb
sentence-transformers
flask
ollama
openai
google-generativeai
pypdf
python-dotenv
pandas
openpyxl
